Fixed missing API for VisualC reflected binary
For some C++ classes, generic parameters doesn't exist before `Resolve` is called. Added `type = type.Resolve();` for types with IsRequiredModifier == true.

Closes #200
