+++ mdoc/mdoc.Test/BasicFormatterTests.cs
@@ -10,20 +10,6 @@ namespace mdoc.Test
     {
         protected abstract T formatter { get; }
 
-        protected MethodDefinition GetMethod(Type type, Func<MethodDefinition, bool> query)
-        {
-            return GetMethod(GetType(type), query);
-        }
-
-        protected MethodDefinition GetMethod(TypeDefinition testclass, Func<MethodDefinition, bool> query)
-        {
-            var methods = testclass.Methods;
-            var member = methods.FirstOrDefault(query)?.Resolve();
-            if (member == null)
-                throw new Exception("Did not find the member in the test class");
-            return member;
-        }
-
         protected MethodDefinition GetProperty(TypeDefinition testclass, Func<MethodDefinition, bool> query)
         {
             var methods = testclass.Methods;
