+++ mdoc/mdoc.Test/MDocUpdaterTests.cs
@@ -1,8 +1,10 @@
 ﻿using System.Collections.Generic;
 using Mono.Cecil;
+using Mono.Cecil.Rocks;
 using Mono.Collections.Generic;
 using Mono.Documentation;
 using NUnit.Framework;
+using Cpp = Mono_DocTest_Generic;
 
 namespace mdoc.Test
 {
@@ -21,6 +23,16 @@ namespace mdoc.Test
 
             Assert.AreEqual(1, attributes.Count);
             Assert.IsEmpty(customAttributes);
+        }
+
+        [Test]
+        public void Test_GetDocParameterType_CppGenericParameterType_ReturnsTypeWithGenericParameters()
+        {
+            var method = GetMethod(typeof(Cpp.GenericBase<>), "BaseMethod2");
+
+            string parameterType = MDocUpdater.GetDocParameterType(method.Parameters[0].ParameterType);
+
+            Assert.AreEqual("Mono_DocTest_Generic.GenericBase<U>", parameterType);
         }
     }
 }
